Simplify tests using Policy.read_json_reform method
