+++ taxcalc/tests/test_calculator.py
@@ -383,116 +383,6 @@ def test_calculator_using_nonstd_input():
     assert np.allclose(mtr_ptax, exp_mtr_ptax)
 
 
-REFORM_JSON = """
-// Example of a reform file suitable for read_json_param_objects().
-// This JSON file can contain any number of trailing //-style comments, which
-// will be removed before the contents are converted from JSON to a dictionary.
-// Within each "policy" object, the primary keys are parameters and
-// the secondary keys are years.
-// Both the primary and secondary key values must be enclosed in quotes (").
-// Boolean variables are specified as true or false (no quotes; all lowercase).
-{
-  "policy": {
-    "AMT_brk1": // top of first AMT tax bracket
-    {"2015": 200000,
-     "2017": 300000
-    },
-    "EITC_c": // maximum EITC amount by number of qualifying kids (0,1,2,3+)
-    {"2016": [ 900, 5000,  8000,  9000],
-     "2019": [1200, 7000, 10000, 12000]
-    },
-    "II_em": // personal exemption amount (see indexing changes below)
-    {"2016": 6000,
-     "2018": 7500,
-     "2020": 9000
-    },
-    "II_em-indexed": // personal exemption amount indexing status
-    {"2016": false, // values in future years are same as this year value
-     "2018": true   // values in future years indexed with this year as base
-    },
-    "SS_Earnings_c": // social security (OASDI) maximum taxable earnings
-    {"2016": 300000,
-     "2018": 500000,
-     "2020": 700000
-    },
-    "AMT_em-indexed": // AMT exemption amount indexing status
-    {"2017": false, // values in future years are same as this year value
-     "2020": true   // values in future years indexed with this year as base
-    }
-  }
-}
-"""
-
-
-@pytest.mark.parametrize("set_year", [False, True])
-def test_read_json_reform_file_and_implement_reform(set_year):
-    """
-    Test reading and translation of reform JSON into a reform dictionary
-    that is then used to call implement_reform method and Calculate.calc_all()
-    NOTE: implement_reform called when policy.current_year == policy.start_year
-    """
-    pol = Policy()
-    if set_year:
-        pol.set_year(2015)
-    param_dict = Calculator.read_json_param_objects(REFORM_JSON, None)
-    pol.implement_reform(param_dict['policy'])
-    syr = pol.start_year
-    # pylint: disable=protected-access,no-member
-    amt_brk1 = pol._AMT_brk1
-    assert amt_brk1[2015 - syr] == 200000
-    assert amt_brk1[2016 - syr] > 200000
-    assert amt_brk1[2017 - syr] == 300000
-    assert amt_brk1[2018 - syr] > 300000
-    ii_em = pol._II_em
-    assert ii_em[2016 - syr] == 6000
-    assert ii_em[2017 - syr] == 6000
-    assert ii_em[2018 - syr] == 7500
-    assert ii_em[2019 - syr] > 7500
-    assert ii_em[2020 - syr] == 9000
-    assert ii_em[2021 - syr] > 9000
-    amt_em = pol._AMT_em
-    assert amt_em[2016 - syr, 0] > amt_em[2015 - syr, 0]
-    assert amt_em[2017 - syr, 0] > amt_em[2016 - syr, 0]
-    assert amt_em[2018 - syr, 0] == amt_em[2017 - syr, 0]
-    assert amt_em[2019 - syr, 0] == amt_em[2017 - syr, 0]
-    assert amt_em[2020 - syr, 0] == amt_em[2017 - syr, 0]
-    assert amt_em[2021 - syr, 0] > amt_em[2020 - syr, 0]
-    assert amt_em[2022 - syr, 0] > amt_em[2021 - syr, 0]
-    add4aged = pol._ID_Medical_frt_add4aged
-    assert add4aged[2015 - syr] == -0.025
-    assert add4aged[2016 - syr] == -0.025
-    assert add4aged[2017 - syr] == 0.0
-    assert add4aged[2022 - syr] == 0.0
-
-
-def test_json_reform_url():
-    """
-    Test reading a JSON reform from a URL. Results from the URL are expected
-    to match the results from the string.
-    """
-    reform_str = """
-    {
-    "policy": {
-        // raise FICA payroll tax rate in 2018 and 2020
-        "FICA_ss_trt": {
-            "2018": 0.130,
-            "2020": 0.140
-        },
-        // raise Medicare payroll tax rate in 2019 and 2021
-        "FICA_mc_trt": {
-            "2019": 0.030,
-            "2021": 0.032
-        }
-      }
-    }
-    """
-    reform_url = ('https://raw.githubusercontent.com/PSLmodels/'
-                  'Tax-Calculator/master/taxcalc/reforms/ptaxes0.json')
-    params_str = Calculator.read_json_param_objects(reform_str, None)
-    params_url = Calculator.read_json_param_objects(reform_url, None)
-    assert params_str == params_url
-
-
 def test_bad_json_names(tests_path):
     """
     Test that ValueError raised with assump or reform do not end in '.json'
@@ -506,37 +396,6 @@ def test_bad_json_names(tests_path):
         Calculator.read_json_param_objects(None, csvname)
     with pytest.raises(ValueError):
         Calculator.read_json_param_objects(None, 'http://name.json.html')
-
-
-def test_read_bad_json_reform_file():
-    """
-    Test invalid JSON reform files.
-    """
-    badreform1 = """
-    {
-      "policy": { // example of incorrect JSON because 'x' must be "x"
-        'x': {"2014": 4000}
-      }
-    }
-    """
-    badreform2 = """
-    {
-      "title": "",
-      "policyx": { // example of reform file not containing "policy" key
-        "SS_Earnings_c": {"2018": 9e99}
-      }
-    }
-    """
-    with pytest.raises(ValueError):
-        Calculator.read_json_param_objects(badreform1, None)
-    with pytest.raises(ValueError):
-        Calculator.read_json_param_objects(badreform2, None)
-    with pytest.raises(ValueError):
-        Calculator.read_json_param_objects(list(), None)
-    with pytest.raises(ValueError):
-        Calculator.read_json_param_objects(None, 'unknown_file_name')
-    with pytest.raises(ValueError):
-        Calculator.read_json_param_objects(None, list())
 
 
 def test_json_assump_url():
@@ -656,8 +515,6 @@ def test_calc_all():
     """
     cyr = 2016
     pol = Policy()
-    param_dict = Calculator.read_json_param_objects(REFORM_JSON, None)
-    pol.implement_reform(param_dict['policy'])
     pol.set_year(cyr)
     nonstd = Records(data=pd.read_csv(StringIO(RAWINPUT_CONTENTS)),
                      start_year=cyr, gfactors=None, weights=None)
@@ -665,7 +522,6 @@ def test_calc_all():
     calc = Calculator(policy=pol, records=nonstd,
                       sync_years=False)  # keeps raw data unchanged
     assert calc.current_year == cyr
-    assert calc.reform_warnings == ''
 
 
 def test_noreform_documentation():
@@ -674,7 +530,6 @@ def test_noreform_documentation():
     """
     reform_json = """
     {
-    "policy": {}
     }
     """
     assump_json = """
@@ -705,7 +560,6 @@ def test_reform_documentation():
     """
     reform_json = """
 {
-"policy": {
     "II_em-indexed": {
         "2016": false,
         "2018": true
@@ -728,7 +582,6 @@ def test_reform_documentation():
     "ID_BenefitCap_Switch": {
         "2020": [false, false, false, false, false, false, false]
     }
-}
 }
 """
     assump_json = """
