+++ taxcalc/tests/test_puf_var_stats.py
@@ -151,9 +151,9 @@ def test_puf_var_stats(tests_path, puf_fullsample):
     """
     # create a baseline Policy object containing 2017_law.json parameters
     pre_tcja_jrf = os.path.join(tests_path, '..', 'reforms', '2017_law.json')
-    pre_tcja = Calculator.read_json_param_objects(pre_tcja_jrf, None)
+    pre_tcja = Policy.read_json_reform(pre_tcja_jrf)
     baseline_policy = Policy()
-    baseline_policy.implement_reform(pre_tcja['policy'])
+    baseline_policy.implement_reform(pre_tcja)
     # create a Calculator object using baseline_policy and full puf.csv sample
     rec = Records(data=puf_fullsample)
     calc = Calculator(policy=baseline_policy, records=rec, verbose=False)
